requirements.txt
appdirs==1.4.3
argcomplete==1.9.3
attrs==19.1.0
black==19.3b0
#boto3==1.9.228
#botocore==1.12.228
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
docutils==0.15.2
durationpy==0.5
Flask==2.2.2
#Flask-S3==0.3.3
future==0.16.0
hjson==3.0.1
idna==2.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.9.3
kappa==0.6.0
lambda_packages==0.20.0
MarkupSafe==3.0.2
pillow==11.0.0
placebo==0.9.0
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==5.1.2
requests==2.32.3
#s3transfer==0.2.1
six==1.12.0
toml==0.10.0
tqdm==4.19.1
Unidecode==1.1.1
urllib3==2.2.3
Werkzeug==2.2.2
wsgi-request-logger==0.4.6
